[refacotr] #42 - home 목 데이터 분리
