+++ src/components/home/body/ReminderCalender.tsx
@@ -3,37 +3,11 @@ import LargeView from '@svg/large-view-icon.svg';
 import SmallView from '@svg/small-view-icon.svg';
 import { useEffect, useState } from 'react';
 // import Calender from './Calender';
-import { CalenderPostResponse } from '@/types/reminder';
 import InsightList from './InsightList';
 import dayjs from 'dayjs';
 import CalenderModal from './CalenderModal';
 import Calender2 from './Calender2';
-
-export const calenderData: CalenderPostResponse = {
-  date: '2024-03-18',
-  remindRead: 1,
-  remindTotal: 2,
-  remindInsightList: [
-    {
-      insightId: 2,
-      insightMainImage: '/image/디자인1.jpg',
-      insightTitle: '알아두면 쓸모있는 시멘틱 마크업 개념',
-      insightSummary:
-        '미드저니는 UX/UI디자인, 그래픽 디자인 등 다양한 분야에서 활용될 수있습니다. 미드저니를 활용해 UX/UI 디자인을 수행하는 경우, 시나리오와 퍼소나를 아주 높은 퀄리티로 시각화 할 수 있습니다.',
-      insightTagList: ['UI/UX', '사용자 경험'],
-      todayRead: false,
-    },
-    {
-      insightId: 2,
-      insightMainImage: '/image/개발1.jpg',
-      insightTitle: '디자인시스템에 모션 가이드 추가하는 방법',
-      insightSummary:
-        '미드저니는 UX/UI디자인, 그래픽 디자인 등 다양한 분야에서 활용될 수있습니다. 미드저니를 활용해 UX/UI 디자인을 수행하는 경우, 시나리오와 퍼소나를 아주 높은 퀄리티로 시각화 할 수 있습니다.',
-      insightTagList: ['UI/UX', '사용자 경험'],
-      todayRead: true,
-    },
-  ],
-};
+import { calenderData } from '@/constants/data';
 
 // TODO [2] - 날짜 클릭 시 해당 날짜에 리마인드 해야 하는 인사이트 호출
 const ReminderCalender = () => {
